+++ src/types/forms/uploadStorageBucketBackupFile.d.ts
@@ -1,0 +1,6 @@
+export interface UploadStorageBucketBackupFileFormValues {
+  bucketFile: File | null;
+  name: string;
+  pool: string;
+  clusterMember: string;
+}